Introduce abstraction to communicate with platform
