--- lib/src/web/bitmovin_player_method_channel.dart
@@ -1,17 +1,0 @@
-import 'package:bitmovin_player/src/web/bitmovin_player_platform_interface.dart';
-import 'package:flutter/foundation.dart';
-import 'package:flutter/services.dart';
-
-/// An implementation of [BitmovinPlayerPlatform] that uses method channels.
-class MethodChannelBitmovinPlayer extends BitmovinPlayerPlatform {
-  /// The method channel used to interact with the native platform.
-  @visibleForTesting
-  final methodChannel = const MethodChannel('bitmovin_player');
-
-  @override
-  Future<String?> getPlatformVersion() async {
-    final version =
-        await methodChannel.invokeMethod<String>('getPlatformVersion');
-    return version;
-  }
-}
